requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-jalali==6.0.0
django-jalali-date==1.0.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
jdatetime==4.0.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5